requirements.txt
loguru
PyYAML
toml